requirements.txt
gspread==6.1.2
google-auth
playwright
python-dotenv
thefuzz
python-levenshtein
jinja2
